Reduce package size
Some files are for development only and don't need to be included
production releases. Package Control downloads the archive from Github
which exports the repository. When Git exports a repository it uses
configurations in `.gitattributes`. Files and directories with the
attribute `export-ignore` won’t be added to archive files.

See https://git-scm.com/docs/gitattributes

There are probably some other files and directories that can be safely
ignored in releases e.g. the sources directory, but I don't know
enough about the package to know if they can be ignored safely.